@@ -5,7 +5,7 @@
 
     <groupId>at.co.brandstetter</groupId>
     <artifactId>aircontrol</artifactId>
-    <version>0.0.6-SNAPSHOT</version>
+    <version>0.7.0-SNAPSHOT</version>
     <packaging>jar</packaging>
     <name>aircontrol</name>
     <description>Control service for Drexel &amp; Weiss ventilation systems</description>
@@ -13,12 +13,12 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.7.13</version>
+        <version>3.4.3</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
 
     <properties>
-        <java-version>17</java-version>
+        <java-version>22</java-version>
         <start-class>at.co.brandstetter.aircontrol.AircontrolApplication</start-class>
     </properties>
 
@@ -79,26 +79,26 @@
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
-            <version>1.18.28</version>
+            <version>1.18.36</version>
         </dependency>
 
         <dependency>
             <groupId>com.fazecast</groupId>
             <artifactId>jSerialComm</artifactId>
-            <version>2.10.2</version>
+            <version>2.11.0</version>
         </dependency>
 
         <dependency>
             <groupId>io.micrometer</groupId>
             <artifactId>micrometer-core</artifactId>
-            <version>1.11.1</version>
+            <version>1.14.4</version>
         </dependency>
 
         <!-- Using rate limiting service -->
         <dependency>
             <groupId>com.google.guava</groupId>
             <artifactId>guava</artifactId>
-            <version>32.0.0-jre</version>
+            <version>33.4.0-jre</version>
         </dependency>
 
         <!-- Failsafe Plugin -->
@@ -134,8 +134,8 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration>
-                    <source>17</source>
-                    <target>17</target>
+                    <source>22</source>
+                    <target>22</target>
                 </configuration>
             </plugin>
 
