@@ -9,6 +9,7 @@ import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.ApplicationListener;
 import org.springframework.stereotype.Component;
+import javax.annotation.Nonnull;
 
 @SpringBootApplication
 @EnableCaching
@@ -33,7 +34,7 @@ class MyApplicationListener
     }
 
     @Override
-    public void onApplicationEvent(ApplicationReadyEvent event) {
+    public void onApplicationEvent(@Nonnull ApplicationReadyEvent event) {
 
         modbus.open();
 
